Move components to their respective folders
