+++ application/ui/src/features/dataset/media-preview/secondary-toolbar/prediction-inference-devices/prediction-inference-devices.test.tsx
