+++ application/ui/src/features/dataset/media-preview/secondary-toolbar/prediction-model-selector/prediction-model-selector.component.tsx
